backend/requirements-celery.txt
celery==5.2.3
django-celery-beat==2.5.0
daphne==4.0.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
channels==4.0.0
channels-redis==4.1.0
django-extensions==3.2.3
django-rest-passwordreset==1.3.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
better-profanity==0.7.0
djangorestframework-simplejwt==5.3.0
sendgrid==6.11.0
requests==2.31.0
Markdown==3.5.1